+++ main.go
@@ -16,47 +16,62 @@ func main() {
 		log.Printf(
 			"Possible usages:\n" +
 				"  upgrade-local-node-go <target_version>\n" +
-				"  upgrade-local-node-go vote [proposal-id]\n",
+				"  upgrade-local-node-go vote [proposal-id]\n" +
+				"  upgrade-local-node-go deposit [proposal-id]\n",
 		)
 		os.Exit(1)
 	}
 
 	bin, err := utils.NewEvmosTestingBinary()
 	if err != nil {
-		log.Fatalf("Error creating binary: %v", err)
+		log.Fatalf("error creating binary: %v", err)
 	}
 
-	err = bin.GetAccounts()
-	if err != nil {
-		log.Fatalf("Error getting accounts: %v", err)
+	if err = bin.GetAccounts(); err != nil {
+		log.Fatalf("error getting accounts: %v", err)
 	}
 
-	//nolint:nestif // nesting complexity is fine here, will be reworked with Cobra commands anyway
-	if os.Args[1] == "vote" {
-		proposalID, err := getProposalIDForVoting(bin, os.Args)
+	// TODO: use with Cobra CLI
+	switch os.Args[1] {
+	case "vote":
+		proposalID, err := getProposalIDFromInput(bin, os.Args)
 		if err != nil {
-			log.Fatalf("Error getting proposal ID: %v", err)
+			log.Fatalf("error getting proposal ID: %v", err)
 		}
 
-		err = gov.SubmitAllVotesForProposal(bin, proposalID)
-		if err != nil {
-			log.Fatalf("Error submitting votes for proposal %d: %v", proposalID, err)
+		if err = gov.SubmitAllVotesForProposal(bin, proposalID); err != nil {
+			log.Fatalf("error submitting votes for proposal %d: %v", proposalID, err)
 		}
-	} else {
+
+	case "deposit":
+		deposit, err := gov.GetMinDeposit(bin)
+		if err != nil {
+			log.Fatalf("error getting minimum deposit: %v", err)
+		}
+
+		proposalID, err := getProposalIDFromInput(bin, os.Args)
+		if err != nil {
+			log.Fatalf("error getting proposal ID: %v", err)
+		}
+
+		if _, err = gov.DepositForProposal(bin, proposalID, bin.Accounts[0].Name, deposit.String()); err != nil {
+			log.Fatalf("error depositing for proposal %d: %v", proposalID, err)
+		}
+
+	default:
 		targetVersion := os.Args[1]
 		if matched, _ := regexp.MatchString(`v\d+\.\d+\.\d(-rc\d+)?`, targetVersion); !matched {
-			log.Fatalf("Invalid target version: %s. Please use the format vX.Y.Z(-rc*).\n", targetVersion)
+			log.Fatalf("invalid target version: %s; please use the format vX.Y.Z(-rc*).\n", targetVersion)
 		}
 
-		err := upgradeLocalNode(bin, targetVersion)
-		if err != nil {
-			log.Fatalf("Error upgrading local node: %v", err)
+		if err := upgradeLocalNode(bin, targetVersion); err != nil {
+			log.Fatalf("error upgrading local node: %v", err)
 		}
 	}
 }
 
-// getProposalIDForVoting gets the proposal ID from the command line arguments.
-func getProposalIDForVoting(bin *utils.Binary, args []string) (int, error) {
+// getProposalIDFromInput gets the proposal ID from the command line arguments.
+func getProposalIDFromInput(bin *utils.Binary, args []string) (int, error) {
 	var (
 		err        error
 		proposalID int
@@ -66,15 +81,15 @@ func getProposalIDForVoting(bin *utils.Binary, args []string) (int, error) {
 	case 2:
 		proposalID, err = gov.QueryLatestProposalID(bin)
 		if err != nil {
-			return 0, errors.Wrap(err, "Error querying latest proposal ID")
+			return 0, errors.Wrap(err, "error querying latest proposal ID")
 		}
 	case 3:
 		proposalID, err = strconv.Atoi(args[2])
 		if err != nil {
-			return 0, errors.Wrapf(err, "Error converting proposal ID %s to integer", args[2])
+			return 0, errors.Wrapf(err, "error converting proposal ID %s to integer", args[2])
 		}
 	default:
-		return 0, errors.New("Invalid number of arguments")
+		return 0, errors.New("invalid number of arguments")
 	}
 
 	return proposalID, nil
@@ -85,7 +100,7 @@ func getProposalIDForVoting(bin *utils.Binary, args []string) (int, error) {
 func upgradeLocalNode(bin *utils.Binary, targetVersion string) error {
 	currentHeight, err := utils.GetCurrentHeight(bin)
 	if err != nil {
-		return errors.Wrap(err, "Error getting current height")
+		return errors.Wrap(err, "error getting current height")
 	}
 
 	upgradeHeight := currentHeight + utils.DeltaHeight
@@ -94,14 +109,13 @@ func upgradeLocalNode(bin *utils.Binary, targetVersion string) error {
 
 	proposalID, err := gov.SubmitUpgradeProposal(bin, targetVersion, upgradeHeight)
 	if err != nil {
-		return errors.Wrap(err, "Error executing upgrade proposal")
+		return errors.Wrap(err, "error executing upgrade proposal")
 	}
 
 	log.Printf("Scheduled upgrade to %s at height %d.\n", targetVersion, upgradeHeight)
 
-	err = gov.SubmitAllVotesForProposal(bin, proposalID)
-	if err != nil {
-		return errors.Wrapf(err, "Error submitting votes for proposal %d", proposalID)
+	if err = gov.SubmitAllVotesForProposal(bin, proposalID); err != nil {
+		return errors.Wrapf(err, "error submitting votes for proposal %d", proposalID)
 	}
 
 	return nil